requirements.txt
alembic==1.6.5
appdirs==1.4.4
astroid==2.5.6
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
boto3==1.17.110
botocore==1.20.110
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
decorator==5.0.7
distlib==0.3.1
dnspython==2.1.0
docutils==0.17.1
email-validator==1.1.3
filelock==3.0.12
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==2.10
iniconfig==1.1.1
instaloader==4.7.1
ipython==7.23.1
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
MouseInfo==0.1.3
mysql-connector-python==8.0.25
numpy==1.20.2
packaging==21.0
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
pluggy==0.13.1
prompt-toolkit==3.0.18
protobuf==3.17.0
py==1.10.0
PyAutoGUI==0.9.52
pycparser==2.20
pycryptodome==3.10.1
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.8.1
pylint==2.8.2
PyMsgBox==1.0.9
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyRect==0.1.4
PyScreeze==0.1.27
pytest==6.2.4
python-dateutil==2.8.1
python-editor==1.0.4
PythonTurtle==0.3.2
PyTweening==1.0.3
pytz==2021.1
pywin32==300
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.4.2
secret==0.8
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.21
tabulate==0.8.9
tk==0.1.0
toml==0.10.2
traitlets==5.0.5
turtle==0.0.1
urllib3==1.26.4
virtualenv==20.4.6
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
wtf==0.1
WTForms==2.3.3
